+++ raja/raja_driver.hpp
@@ -77,7 +77,7 @@ class RajaDriver : public Driver {
         umpire::Allocator device_allocator;
 
         int sx, sy, sz;
-        int dx, dy, dz, dt;
+        float dx, dy, dz, dt;
         int nx, ny, nz;
         int bord, absorb;
         host_data_container   host_data;
